Factor _locate_parameters into AbstractNetworkValidator
Move the shared parameter-lookup logic into the base class so
both PydanticParametersNetworkValidator and
SemanticParametersNetworkValidator inherit it without
duplication.

Also removes _resolve_agent_name (agent_name from
get_name_to_spec is sufficient) and adds an inline comment
explaining the string-items commondef substitution.
